Progress

+++ Software/ESP32/main/calibrate.h
@@ -7,8 +7,8 @@
 /*
  * Functions
  */
-void  calibrate(void);
-float solve_spline(float theta);  // Use the spline coeficients to find the scale factor
-void  get_calibration_data(void); // Retrieve calibration data from NONVOL
-void  calibration_test(void);     // Generate test calibration data
+void   calibrate(void);
+double solve_spline(double theta); // Use the spline coeficients to find the scale factor
+void   get_calibration_data(void); // Retrieve calibration data from NONVOL
+void   calibration_test(void);     // Generate test calibration data
 #endif
